requirements.txt
google-genai
Pillow
python-dotenv
Flask
torch
numpy
transformers
pandas